LayoutTests/http/tests/security/contentSecurityPolicy/strict-dynamic-mixed-order-scheme-list-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/simpleSourcedScript.js because it does not appear in the script-src directive of the Content Security Policy.

PASS All the expected CSP violation reports have been fired.
